AppWindow: Use CommandManager directly. View: Change some function return types from int to unsigned int

+++ src/application/view.cpp
@@ -102,9 +102,9 @@ void View::clearSelection()
     selectedPagesChanged.emit();
 }
 
-int View::getSelectedChildIndex() const
+unsigned int View::getSelectedChildIndex() const
 {
-    return static_cast<int>(getSelectedChildrenIndexes().front());
+    return getSelectedChildrenIndexes().front();
 }
 
 std::vector<unsigned int> View::getSelectedChildrenIndexes() const
